Move to 16 bit waveform samples and amplitude. Bugged "envelope"

+++ ATtiny817_Test/synth/oscillator.h
@@ -33,7 +33,7 @@ void osc_init();
 
 void osc_set_waveform(oscillator_t osc, waveform_t waveform);
 void osc_set_frequency(oscillator_t osc, uint16_t frequency_dHz);
-void osc_set_amplitude(oscillator_t osc, uint8_t amplitude);
+void osc_set_amplitude(oscillator_t osc, uint16_t amplitude);
 void osc_set_filter_value(oscillator_t osc, uint8_t filter_value);
 void osc_set_sync(bool enabled);
